+++ changelog.md
@@ -17,6 +17,7 @@ Codion Change Log
 - FilterTableCellRendererFactory moved to FilterTableCellRenderer and renamed Factory.
 - FilterTableCellEditorFactory moved to FilterTableCellEditor and renamed Factory.
 - ColumnConditionPanels.selectablePanels() and selectPanel() renamed selectable() and select() respectively.
+- FilterTable.Builder.conditionState() renamed filterState().
 ### is.codion.framework.domain
 - Entity.copy(), copyBuilder() and deepCopy() moved to Entity.Copy, renamed mutable(), builder() and deep() respectively, immutable() added.
 ### is.codion.framework.db